Backend/requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
pillow
django-cors-headers